requirements.txt
asgiref==3.11.0
attrs==25.4.0
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
cryptography==46.0.3
dj-database-url==3.0.1
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.2
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.4
typing_extensions==4.15.0
uritemplate==4.2.0
whitenoise==6.11.0
